AIOD-223 Implement finetuning workflow with model checkpoint handling and argument parser
